Moved certain tests inside of GlitchLib and you can still run them the same way.
