+++ src/main/java/frc/robot/Robot.java
@@ -4,8 +4,8 @@
 
 package frc.robot;
 
-import Glitch.Controller.Controller;
-import Glitch.NetworkTableLogger;
+import Glitch.Lib.Controller.Controller;
+import Glitch.Lib.NetworkTableLogger;
 
 import com.ctre.phoenix6.swerve.SwerveModule;
 import com.ctre.phoenix6.swerve.SwerveRequest;
